pyproject.toml
[project]
name = "music-sorter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "librosa>=0.10.2.post1",
    "musicbrainzngs>=0.7.1",
    "mutagen>=1.47.0",
    "numpy>=2.2.6",
    "pyacoustid>=1.3.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
